refactored, fixed, and updated the 'required modules automated install' code
